Add offset-addressed GPU ops and the sconv history-roll kernel
Task 9 prep: _off twins of matvec/rmsnorm/sconv/rel_project let the
real-model graph read F32 resident weights (norms, rel_proj, router,
sconv weights) directly from the wrapped resident.bin buffer at their
manifest byte offset instead of a fresh per-call upload. A new
sepia_copy_f32 kernel materializes single-row views and does GPU-side
KV-cache writes without a host round trip; a new
sepia_sconv_hist_roll_f32 kernel moves Task 8's host-side sconv-history
update onto the GPU (pure data movement, no float op) so the real
graph never needs a sync just to keep sconv history correct; a new
sepia_scale_f32 kernel applies the dense-MLP global_scale without a
host round trip either. All purely additive -- existing ops/gates
untouched, gputest/gpucompare/gpuattn/gpuquants all still green.

Also backports the tau-precision fix flagged in Task 8's review carry:
q_scaled must be `(float)((double)q * tau)` matching attn_forward_chunk
exactly, not a plain float*float multiply -- the tiny oracle's gate
never caught this because tau==1.0 at every tested position (where the
two formulas coincide), but Task 9's real model exercises log-scaling
for real.

+++ src/sepia.c
@@ -4028,10 +4028,19 @@ static void gpu_attn_forward_chunk(const Config *cfg, const LayerWeights *lw, La
         kv_lo[t] = kvlo;
         kv_hi[t] = q_pos;
     }
+    /* tau precision (P2 Task 9 review carry): must match attn_forward_chunk's
+     * own formula bit-for-bit -- `(float)((double)q * tau)`, double-precision
+     * intermediate, not a plain float*float multiply. This tiny-oracle gate
+     * never caught the earlier float-only version because have_log_scaling
+     * is never true here (tau==1.0 at every tested position, where the two
+     * formulas coincide exactly); Task 9's real model does exercise
+     * log-scaling (global layers past the floor position), so the gap is
+     * fixed here too, not just in the real-path code that motivated it. */
     float *q_scaled = xmalloc(sizeof(float) * (size_t)T * (size_t)q_dim);
     for (int t = 0; t < T; t++)
         for (int i = 0; i < q_dim; i++)
-            q_scaled[(size_t)t * q_dim + i] = q_normed[(size_t)t * q_dim + i] * tau[t];
+            q_scaled[(size_t)t * q_dim + i] =
+                (float)((double)q_normed[(size_t)t * q_dim + i] * (double)tau[t]);
     free(q_normed);
 
     /* 6. rel-project. */
